[69] - Refactorizado carpeta UserDataManager
